+++ src/main/java/dev/willyelton/crystal_tools/CrystalTools.java
@@ -3,6 +3,7 @@ package dev.willyelton.crystal_tools;
 import dev.willyelton.crystal_tools.block.ModBlocks;
 import dev.willyelton.crystal_tools.command.RegisterCommandEvent;
 import dev.willyelton.crystal_tools.config.CrystalToolsConfig;
+import dev.willyelton.crystal_tools.crafting.ItemDisabledCondition;
 import dev.willyelton.crystal_tools.crafting.ModRecipes;
 import dev.willyelton.crystal_tools.item.ModItems;
 import dev.willyelton.crystal_tools.item.armor.ModArmor;
@@ -10,8 +11,11 @@ import dev.willyelton.crystal_tools.keybinding.KeyBindings;
 import dev.willyelton.crystal_tools.network.PacketHandler;
 import dev.willyelton.crystal_tools.item.tool.ModTools;
 import dev.willyelton.crystal_tools.world.ModOres;
+import net.minecraft.client.renderer.ItemBlockRenderTypes;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.world.level.block.Block;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.common.crafting.CraftingHelper;
 import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -33,6 +37,7 @@ public class CrystalTools {
     // Directly reference a log4j logger.
     private static final Logger LOGGER = LogManager.getLogger();
     public static final String MODID = "crystal_tools";
+    public static final ItemDisabledCondition.Serializer INSTANCE = new ItemDisabledCondition.Serializer();
 
     public CrystalTools() {
         IEventBus eventBus = FMLJavaModLoadingContext.get().getModEventBus();
@@ -76,11 +81,16 @@ public class CrystalTools {
 
     private void clientSetup(final FMLClientSetupEvent event) {
         KeyBindings.init();
+        ItemBlockRenderTypes.setRenderLayer(ModBlocks.CRYSTAL_TORCH.get(), RenderType.cutout());
+        ItemBlockRenderTypes.setRenderLayer(ModBlocks.CRYSTAL_WALL_TORCH.get(), RenderType.cutout());
     }
 
     private void init(final FMLCommonSetupEvent event) {
         // some preinit code
         event.enqueueWork(ModOres::registerConfiguredFeatures);
+        event.enqueueWork(() -> {
+            CraftingHelper.register(INSTANCE);
+        });
     }
 
     private void setup() {