+++ src/main/java/dev/willyelton/crystal_tools/item/armor/LevelableArmor.java
@@ -3,6 +3,7 @@ package dev.willyelton.crystal_tools.item.armor;
 import com.google.common.collect.ImmutableMultimap;
 import com.google.common.collect.Multimap;
 import dev.willyelton.crystal_tools.CreativeTabs;
+import dev.willyelton.crystal_tools.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.item.LevelableItem;
 import dev.willyelton.crystal_tools.item.ModItems;
 import dev.willyelton.crystal_tools.utils.LevelUtils;
@@ -143,8 +144,34 @@ public class LevelableArmor extends ArmorItem implements LevelableItem, Wearable
 
     @Override
     public void onArmorTick(ItemStack stack, Level world, Player player) {
+        if (this.isDisabled()) {
+            stack.shrink(1);
+            return;
+        }
+
         if (!ToolUtils.isBroken(stack) && NBTUtils.getFloatOrAddKey(stack, "night_vision") > 0) {
             player.addEffect(new MobEffectInstance(MobEffects.NIGHT_VISION, 400, 0, false, false));
+        }
+    }
+
+    @Override
+    public boolean isDisabled() {
+        switch (this.itemType) {
+            case "helmet" -> {
+                return CrystalToolsConfig.DISABLE_HELMET.get();
+            }
+            case "chestplate" -> {
+                return CrystalToolsConfig.DISABLE_CHESTPLATE.get();
+            }
+            case "leggings" -> {
+                return CrystalToolsConfig.DISABLE_LEGGINGS.get();
+            }
+            case "boots" -> {
+                return CrystalToolsConfig.DISABLE_BOOTS.get();
+            }
+            default -> {
+                return false;
+            }
         }
     }
 }