+++ src/main/java/dev/willyelton/crystal_tools/item/tool/DiggerLevelableTool.java
@@ -20,7 +20,7 @@ import org.jetbrains.annotations.NotNull;
  * Class for the shovel and pickaxe
  * Handles the 3x3 mining
  */
-public class DiggerLevelableTool extends LevelableTool {
+public abstract class DiggerLevelableTool extends LevelableTool {
     private static int MAX_VEIN_MINER_DEPTH = 4;
 
     public DiggerLevelableTool(Properties properties, TagKey<Block> mineableBlocks, String toolType) {