+++ src/main/java/dev/willyelton/crystal_tools/item/tool/PickaxeLevelableTool.java
@@ -1,6 +1,8 @@
 package dev.willyelton.crystal_tools.item.tool;
 
+import dev.willyelton.crystal_tools.config.CrystalToolsConfig;
 import dev.willyelton.crystal_tools.utils.NBTUtils;
+import dev.willyelton.crystal_tools.utils.ToolUseUtils;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.tags.BlockTags;
@@ -20,40 +22,16 @@ public class PickaxeLevelableTool extends DiggerLevelableTool {
     }
 
     public InteractionResult useOn(UseOnContext context) {
-        ItemStack tool = context.getItemInHand();
-
-        if (NBTUtils.getFloatOrAddKey(tool, "torch") > 0) {
-            Level level = context.getLevel();
-            BlockPos position = context.getClickedPos();
-            BlockState state = level.getBlockState(position);
-            Direction direction = context.getClickedFace();
-            position = position.relative(direction);
-
-            BlockState torchBlockState;
-
-            if (direction.equals(Direction.UP)) {
-                torchBlockState = Blocks.TORCH.defaultBlockState();
-            } else if (direction.equals(Direction.DOWN)) {
-                return InteractionResult.PASS;
-            } else {
-                torchBlockState = Blocks.WALL_TORCH.defaultBlockState().setValue(WallTorchBlock.FACING, direction);
-            }
-
-            if (level.getBlockState(position).is(Blocks.AIR) && state.isFaceSturdy(level, position, direction)) {
-                level.setBlock(position, torchBlockState, 0);
-            } else {
-                return InteractionResult.FAIL;
-            }
-
-            if (!level.isClientSide && context.getPlayer() != null) {
-                tool.hurtAndBreak(10, context.getPlayer(), (player) -> {
-                    player.broadcastBreakEvent(EquipmentSlot.MAINHAND);
-                });
-            }
-        } else {
-            return super.useOn(context);
+        if (this.isDisabled()) {
+            context.getItemInHand().shrink(1);
+            return InteractionResult.FAIL;
         }
 
-        return InteractionResult.FAIL;
+        return ToolUseUtils.useOnTorch(context, this);
+    }
+
+    @Override
+    public boolean isDisabled() {
+        return CrystalToolsConfig.DISABLE_PICKAXE.get();
     }
 }